requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
atlastk==0.13.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
decouple==0.0.7
dictdiffer==0.9.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-pagination==0.12.19
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.4
iso8601==1.1.0
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
pbr==6.0.0
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
Tempita==0.5.2
tomlkit==0.12.4
tortoise==0.1.1
tortoise-orm==0.20.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
